Backend/mealapp/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow

Flask==2.0.1
itsdangerous==2.0.1
Flask-Mail==0.9.1

